requirements.txt
Django==1.11.29
djangorestframework==3.9.4
Pygments==2.5.2
pytz==2020.1
